fastf1_openf1_audit_toolkit/requirements.txt
fastf1>=3.8.1
pandas>=2.2.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.18
python-dotenv>=1.0.1
pyyaml>=6.0.1
